+++ ft_memcpy.c
@@ -6,7 +6,7 @@
 /*   By: mstupnik <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/09/18 10:22:48 by mstupnik          #+#    #+#             */
-/*   Updated: 2019/09/19 13:39:03 by mstupnik         ###   ########.fr       */
+/*   Updated: 2019/09/19 14:35:09 by mstupnik         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,7 +16,7 @@ void	*ft_memcpy(void *dst, const void *src, size_t n)
 {
 	char	*cdst;
 	char	*csrc;
-	int		i;
+	size_t	i;
 
 	cdst = (char *)dst;
 	csrc = (char *)src;
